add all unit test for wishlist services

+++ src/main/java/org/shopby_backend/wishlist/dto/WishListGetAllByIdDto.java
@@ -1,4 +1,4 @@
 package org.shopby_backend.wishlist.dto;
 
-public record WishListGetAllByIdDto(Integer userId) {
+public record WishListGetAllByIdDto(Long userId) {
 }
