+++ src/main/java/org/shopby_backend/wishlist/persistence/WishlistRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface WishlistRepository extends JpaRepository<WishlistEntity, Integer> {
-    List<WishlistEntity> findByUserId(Integer userId);
+    List<WishlistEntity> findByUserId(Long userId);
 }
